+++ table-service/src/main/java/net/kalaha/table/api/GetTableRequest.java
@@ -6,20 +6,20 @@ public class GetTableRequest extends TableRequestAction implements Serializable 
 
 	private static final long serialVersionUID = 2137099430705156440L;
 	
-	private int gameId;
+	private long gameId;
 	
 	public GetTableRequest() { }
 	
-	public GetTableRequest(int userId, int correlationId, int gameId) {
+	public GetTableRequest(long userId, int correlationId, long gameId) {
 		super(userId, correlationId);
 		this.gameId = gameId;
 	}
 	
-	public int getGameId() {
+	public long getGameId() {
 		return gameId;
 	}
 	
-	public void setGameId(int gameId) {
+	public void setGameId(long gameId) {
 		this.gameId = gameId;
 	}
 }
